common: use statement exprs for MPMAX, MPMIN, and MPCLAMP
Prevents double evaluations from occurring.

+++ video/mp_image.c
@@ -663,7 +663,7 @@ void mp_image_clear(struct mp_image *img, int x0, int y0, int x1, int y1)
             int plane_bits = fmt->bpp[cd->plane] * misery;
             if (plane_bits <= 64 && plane_bits % 8u == 0 && cd->size) {
                 plane_size[cd->plane] = plane_bits / 8u;
-                int depth = cd->size + MPMIN(cd->pad, 0);
+                int depth = cd->size + MPMIN((int8_t)cd->pad, 0);
                 double m, o;
                 mp_get_csp_uint_mul(area.params.repr.sys,
                                     area.params.repr.levels,
